Improved benchmarks and the standard global allocator bottleneck
Key Finding: The algorithm for masstree uses allocation patterns that
outperform similar data structures in becnhmarks consistently when
mimalloc is used as the global allocator.

The standard global allocator works best when using SkipList (and will
probably be the best similar concurrent data structure to use if limited
to only the standard global allocator, but various it's not uncommon to
use custom global allocator like jemmaloc and mimalloc, as they were
developed by meta and microsoft respectively for production use).